agents/pyproject.toml
[project]
name = "seraph-agents"
version = "1.0.0"
requires-python = ">=3.12,<4"
readme = "README.md"
license-files = ["LICENSE"]
authors = [{ name = "Seraph" }]

dependencies = [
  # API
  "fastapi[standard]",
  "sse-starlette",
  "pydantic-settings",
  # Models
  "agentscope",
  "openai",
  # Database
  "alembic",
  "pgvector",
  "psycopg[binary]",
  "sqlalchemy",
  "aiosqlite",
  "duckdb",
  # Tools
  "mcp",
  "reportlab",
  "pypdf",
  "duckduckgo-search",
  # Messaging
  "nats-py",
  # Serialization
  "fastavro",
  # Cache
  "cachetools",
  # Tracing
  "opentelemetry-api",
  "opentelemetry-sdk",
  "opentelemetry-instrumentation-fastapi",
]

[project.optional-dependencies]
dev = ["mypy", "ruff", "pytest", "pytest-asyncio"]

[dependency-groups]
dev = [
    "mypy",
    "ruff",
    "pytest",
    "pytest-asyncio",
    "types-cachetools>=6.2.0.20260408",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]

[tool.ruff]
line-length = 120
exclude = [".venv*"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]

[tool.mypy]
check_untyped_defs = true
no_implicit_optional = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]
exclude = [".venv*"]

[[tool.mypy.overrides]]
module = ["pgvector.*", "setuptools.*", "agentscope.*", "duckdb.*"]
ignore_missing_imports = true

[tool.uv.pip]
no-annotate = true

[tool.uv]
default-groups = ["dev"]
